timeout

+++ radio/src/telemetry/telemetry.h
@@ -39,7 +39,7 @@
   #include "flysky_ibus.h"
 #endif
 
-extern uint8_t telemetryStreaming; // >0 (true) == data is streaming in. 0 = no data detected for some time
+extern uint16_t telemetryStreaming; // >0 (true) == data is streaming in. 0 = no data detected for some time
 
 inline bool TELEMETRY_STREAMING()
 {
@@ -55,7 +55,7 @@ extern uint8_t telemetryState;
 
 // FIXME: Hack to make telemetry lost not trigger all the time.
 //constexpr uint8_t TELEMETRY_TIMEOUT10ms = 100; // 1 second
-constexpr uint8_t TELEMETRY_TIMEOUT10ms = 300;
+constexpr uint16_t TELEMETRY_TIMEOUT10ms = 300;
 
 #define TELEMETRY_SERIAL_DEFAULT       0
 #define TELEMETRY_SERIAL_8E2           1
